requirements.txt
tkinterdnd2>=0.3.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
PyInstaller>=5.13
